get render configuration file from server using http request
Some times on worker side is not possible to have a shared render_settings folder and this means that this settings will not be available for rendering jobs. This is solved by sending the content of config file to the worker and save hem in a temporary file for render time usage.